[TAPE] Adding a reversible backprop tape (#815)
* [TAPE] Adds the new QNode which uses the quantum tape

* linting

* Apply suggestions from code review


* Apply suggestions from code review


* docstring fix

* suggested changes

* suggested changes

* suggested changes

* suggested changes

* black

* [TAPE] Adds the qubit parameter-shift tape

* suggested changes

* suggested changes

* suggested changes

* suggested changes

* more

* [TAPE] Adds the CV parameter-shift tape

* more

* everything working

* Apply suggestions from code review


* suggested changes

* fix

* remove observable annotation

* remove observable annotation

* linting

* more

* change to np.allclose

* more changes

* more

* more

* add tests

* bugfix

* finished tests

* docstring

* linting

* fixing tests

* add another test

* another test

* adding reversible tape

* adding tests

* more

* more

* Revert "Merge branch 'tape-pr-6' into tape-pr-7"

This reverts commit 07ed2097df9ddbac1da6d4a87a11908512874531, reversing
changes made to 64dea1e282a622f383d7cf52353cb0e3d49c56f4.

* Revert "Merge branch 'tape-pr-7' into tape-pr-8"

This reverts commit a39d0d11cc39a24d119382973b0d445fe81c6c2c, reversing
changes made to e4530c55b5a8fd8cc279fc74c50ab995b6cafecf.

* Revert "Merge branch 'tape-pr-7' into tape-pr-8"

This reverts commit b16fd58ab65ddf02f145f56dbe4effad63686110, reversing
changes made to 8e0e8d9bd86ff5f165d216b67ca757418b0d2953.

* Apply suggestions from code review


* linting

* black

* black

* suggested changes

* done

* Apply suggestions from code review


* Apply suggestions from code review


* suggested changes

* suggested changes

* Update pennylane/beta/tapes/qnode.py


* black

* suggested changes

+++ pennylane/beta/tapes/tape.py
@@ -1103,7 +1103,7 @@ class QuantumTape(AnnotatedQueue):
         order = options.get("order", 1)
         h = options.get("h", 1e-7)
 
-        shift = np.zeros_like(params)
+        shift = np.zeros_like(params, dtype=np.float64)
         shift[idx] = h
 
         if order == 1:
